#13 [WIP] フロント側へJWTトークンを渡す実装とフロントからemailを受け取る実装の途中
